+++ phpcs.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 use PhpCsFixer\Finder;
 use PhpCsFixer\Config;
@@ -7,9 +7,10 @@ $finder = Finder::create()->in(__DIR__ . '/src');
 
 return Config::create()
     ->setRules([
-        '@PSR2' => true,
+        '@PSR12' => true,
         'array_syntax' => [
             'syntax' => 'short',
         ],
+        'strict_param' => true,
     ])
     ->setFinder($finder);