+++ src/Exception/InvalidCallException.php
@@ -1,12 +1,9 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Bupy7\InputFilter\Exception;
 
 use BadMethodCallException;
 
-/**
- * @author Belosludcev Vasily <https://github.com/bupy7>
- */
-class InvalidCallException extends BadMethodCallException
+final class InvalidCallException extends BadMethodCallException
 {
 }