+++ src/InputFilter/ErrorMessageInterface.php
@@ -1,17 +1,14 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Bupy7\InputFilter\InputFilter;
 
-/**
- * @author Vasily Belosloodcev <https://github.com/bupy7>
- */
 interface ErrorMessageInterface
 {
     /**
      * Set error message for an input and mark an input as invalid.
      * @param string $name
      * @param string $message
-     * @return static
+     * @return ErrorMessageInterface
      */
-    public function setMessage($name, $message);
+    public function setMessage(string $name, string $message): ErrorMessageInterface;
 }