+++ test/Asset/EmptyForm.php
@@ -1,12 +1,13 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Bupy7\InputFilter\Test\Asset;
 
 use Bupy7\InputFilter\FormAbstract;
 
-/**
- * @author Vasily Belosloodcev <https://github.com/bupy7>
- */
-class EmptyForm extends FormAbstract
+final class EmptyForm extends FormAbstract
 {
+    protected function inputs(): array
+    {
+        return [];
+    }
 }