+++ test/Asset/SignInFilter.php
@@ -1,31 +1,28 @@
-<?php
+<?php declare(strict_types=1);
 
 namespace Bupy7\InputFilter\Test\Asset;
 
-use Zend\InputFilter\InputFilter;
-
-/**
- * @author Vasily Belosloodcev <https://github.com/bupy7>
- */
-class SignInFilter extends InputFilter
+final class SignInFilter extends BaseFilter
 {
     public function __construct()
     {
-        $this->add([
-            'name' => 'email',
-            'required' => true,
-            'validators' => [
-                [
-                    'name' => 'EmailAddress',
+        parent::__construct([
+            [
+                'name' => 'email',
+                'required' => true,
+                'validators' => [
+                    [
+                        'name' => 'EmailAddress',
+                    ],
                 ],
             ],
-        ]);
-        $this->add([
-            'name' => 'password',
-            'required' => true,
-            'filters' => [
-                [
-                    'name' => 'StringTrim',
+            [
+                'name' => 'password',
+                'required' => true,
+                'filters' => [
+                    [
+                        'name' => 'StringTrim',
+                    ],
                 ],
             ],
         ]);