+++ CHANGELOG.md
@@ -1,18 +1,29 @@
 php-input-filter
-===============
+===
+
+2.0.0 [2022-01-22]
+---
+
+- Dropped supporting of PHP from 5.6 up to 7.3;
+- Added supporting of PHP up to 8.1;
+- Rewrote Zend packages to Laminas;
+- Dropped `@author` PHPDoc tag;
+- Dropped scenarios;
+- `Bupy7\InputFilter\Exception\NotSupportedException` and `Bupy7\InputFilter\Exception\UnknownPropertyException` is runtime exception now;
+- Added `Bupy7\InputFilter\FormInterface`.
 
 1.0.2 [2020-02-11]
-------------------
+---
 
 - Added support of PHP 7.4.
 
 1.0.1 [2019-02-11]
-------------------
+---
 
 - Code style fixes.
 - Docs fixes.
 
 1.0.0 [2019-02-10]
-------------------
+---
 
 - First release.